added transfers

+++ backend/src/main/java/com/example/bankapp/Account/AccountDto.java
@@ -1,11 +1,11 @@
 package com.example.bankapp.Account;
 
-import jakarta.persistence.Column;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 public class AccountDto {
     private Long number;
-    private Long balance;
+    private BigDecimal balance;
     private LocalDateTime dateOfCreation;
     private String status; // 1 = active, 0 = unactive
 }
